upload

+++ philo/philosophers.h
@@ -6,7 +6,7 @@
 /*   By: dongyeuk <dongyeuk@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/25 14:11:37 by dongyeuk          #+#    #+#             */
-/*   Updated: 2024/02/10 17:21:21 by dongyeuk         ###   ########.fr       */
+/*   Updated: 2024/02/13 16:38:23 by dongyeuk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -54,7 +54,7 @@ typedef struct s_ph
 	pthread_mutex_t	*auth[2];
 	int				tag_no;
 	long long		eat_time;
-	int				meals;
+	long long		meals;
 	t_data			*db;
 }				t_ph;
